Avoid unneccessary database accesses
GORM is performing more database access than is neccessary.
It is overlay aggressive in that any modification to model objects
seems to trigger to insert for the whole object as well as nested
child objects.  This is not necessary, so to avoid, the use of
Omit(clause.Association) is added for all Create/Updates.
db.Debug() shows much less database activity.