Add calculation of principle and interest for Scheduled CashFlows
Principle and Interest payments can be calculated if RepeatInterval
for the ScheduledCashFlow specifies a Rate and if this has Splits
containing associated principle transfer and interest expense.
Works independent of which direction the Transfer payment is going,
allowing the Scheduled CashFlow to be on either side of the
transaction.  Reworked applyRate and updateSplits for code reuse.