Rework tracking of ScheduledCashFlow in applied CashFlows
For CashFlows applied from a ScheduledCashFlow, the lookup for
the ScheduledCashFlow can be simplifed to just store the ID
in CashFlow.RepeatIntervalID as the applied CashFlow doesn't
need the RepeatInterval.  This allows to undo previous fork
of c.Preload function, and c.EditPreload is deleted, and Edits
don't need to load RepeatInterval in this case.
This scheme again matches older behaviour.