Initial routine to get security quotes (using finance-go)
Implement basic fetch routine to get SecurityQuote.
It's not integrated yet as the plan is not to actively get Quotes, but
use a session Cache to store/maintain them.  The point of the cache
will be to limit slamming the database with SecurityValue updates.
Add dependency on finance-go.

Also added here is a helper to calculate business days between 2 dates.
Probably won't be used long term, but was fun to code.
Should limit quotes to once per hour, not once per day.