Complete basic integration of updating Security.Value from Quotes
We don't push any updated security.Values to database yet, but
cached Quotes are used to update securities in List() and so
displayed Securities will be using a recent Price valuation.

Final step is to consider where it makes sense to update security.Value
in database. Should be fine to defer until user logout or session is
terminated (future).
Or just deprecate having it in the database.