Prepare for session support
When supporting sessions, we need to replace GetCurrentUser() with
session.GetCurrentUser() to get active user.  The 'default session'
is still implemented to be used if Sessions are turned off.
But session management now belongs at controller level, so we refactor
to pass down the Session from controller to models and rework usage of
GetCurrentUser() in models.
Mostly the model Create/Get/Update/Delete functions and a few List
functions are updated to pass Session to model, but try to keep use of
Session to just when strictly needed for access control.