Initial implementation of automatic generation of TaxEntries
For Taxes#Index, we add ability to automatically generate Tax Entries,
based on the table of TaxCategories.
These are generated by getting list of CashFlows that match the CategoryID
defined in the TaxCategory.  This includes: Wages, Interest:Mortage,
and Tax Withholding (Taxes:Federal and Taxes:State).
These 'auto' entries are not yet used in the Tax Calculation.

Also fix Taxes#Index to limit to the specified Year.