Add support for Sessions (using scs)
More diffcult that I would like to simply store a pointer to
the allocated model.Session.  We don't want to flatten it as
it contains pointers to maps of cached data.
In the end, it is working as intended, and will allow multi-user
support (signup/login not implemented yet).
A configuration option is added to enable Sessions: global.sessions