Add Account.CashBalance support
Accounts have both a CashBalance and Balance field. Balance is intended
to include value of any Securities (Investment Type) in addition to cash
balance.
Previous neglected to make use of CashBalance, but this is updated now
during updateBalance and used instead of Balance when computing Balances
in ListCashFlows and for the averageDailyBalance calculation.