Update Security.Company#update to not always create new Company
company.Get() will create Company if one with {Name, Symbol) does not
exist.  However, for Updates, it will now first see if can instead update
the Security.Company being updated, for the case when the Company is only
used by that Security and thus no reason to create new one.