Fix Trade#update overwriting Security.Basis
The Trade.Basis was overwriting Trade.Security.Basis.
It seems echo's Bind() will also write form-tagged fields in nested
structures.  Give unique names to the form fields in Security and Company
to prevent this and future problems.
Some controllers are doing explicit Bind of nested structures in some
cases, so perhaps that is unneeded.