Implement TaxEntry generation and TaxItem CashFlows for TradeTypes
For the TaxCategory with .TradeTypeID, now will generate
associated TaxEntry, which provides automatic entries for
Capital Gains and Dividends in taxes#index view. (If
there are any such Trade types for the year.)

And can also provide list of associated Trade CashFlows
(using Trade.toCashFlow()) for the TaxItem#show views.
For this purpose, CashFlow.Amount is the Trade.Gain.

While here, add a link for Sell Trade CashFlows to the gain#show
view.