Fix issue where Date can be wrong in SplitCashFlows
New SplitCashFlows create while processing a Scheduled CashFlow
are likely to have the Date wrong.  The arguments to the goroutine
are passed by value; but for Slices, the underlying elements are
of the slice are shared.  Need to use copy() to make a copy of
the slice and its elements.

Fixes: "Use goroutine to defer inserting any SplitCashFlows"