Fix calculation of interest/principle when applying many repeat cashflows
If tryInsertCashFlow needs to apply more than one CashFlow for a given
ScheduledCashFlow, it needs to update the balance in repeat.Account
so that subsequent interest calculations are using a correct average
daily balance.  Also fix so the balance is carried forward to next
ScheduledCashFlow that might be applied.
While here, fix calculateLoanPI to use the correct Account for finding
principle balance when that is found from the split CashFlow.