Fix Update and Delete of split Transfers
The issue is with the Pair Cashflow for the Transfer.
It was incorrectly determining if it was a Split. If transfer was a
split, the Pair was always being set as a split as well, thus not
updating the Pair account's balance.

While here, since for Transfer Cashflows, we force Date to match in
both CashFlows, do the same for Split transfer.
Meaning the restriction of modifying Date in a split (transfer) is
applied to the pair CashFlow as well.