Update cached AccountBalance after Sell/Buy Trades
Cached Balance is accurate until now saved or overwritten upon
new update of Security Vaules.
Rework updating of AccountBalances to proper test if value
is in the key map to determine if value is cached or not.

GetAccountByName() will return Account now with accurate
Balance information for validation.

test/model/security_test is updated.